+++ src/mortuus.cpp
@@ -686,7 +686,7 @@ struct Mortuus : SanguineModule {
 			}
 		}
 
-		uint8_t buttonsBrightness[apicesCommon::kChannelCount];
+		uint16_t buttonsBrightness[apicesCommon::kChannelCount];
 		for (uint8_t channel = 0; channel < apicesCommon::kChannelCount; ++channel) {
 			switch (processorFunctions[channel]) {
 			case mortuus::FUNCTION_DRUM_GENERATOR:
